Fixed shiftview function
The current implementation is a bit too hacky for my taste, but I
couldn't think of any other way to send two bits of information in
an Arg without splitting it into two functions doing roughly the same
thing.

Also you have to set `arg->i` to something different than 0 when
treating mouse events on the tagbar. Because of this, shiftview uses
the 2nd and 3rd bits of `arg->i` and shiftview calls in `config.h`
are written something like `{.i = (*|*|1)}`.